Remove sourcery pod unused
