--- .sourcery.yml
@@ -1,12 +1,0 @@
-sources:
-  - iOSTemplates/Supports
-templates:
-  - tools/sourcery/templates
-output:
-  iOSTemplatesTests/Sources/Mocks/Sourcery
-args:
-  autoMockableTestableImports:
-    - iOSTemplates
-  autoMockableImports:
-    - RxSwift
-    - RxCocoa
