--- .swiftformat
@@ -1,13 +1,0 @@
-# file options
---exclude Pods, Generated, **/*.generated.swift
-
-# rules
---disable fileHeader
---disable initCoderUnavailable
---disable trailingCommas
---disable wrapEnumCases
---disable wrapMultilineStatementBraces
---disable wrapSwitchCases
---disable blankLinesAtStartOfScope
---stripunusedargs closure-only
---extensionacl on-declarations
