+++ .swiftlint.yml
@@ -25,7 +25,6 @@ opt_in_rules:
   - explicit_init
   - fallthrough
   - fatal_error_message
-  - file_types_order
   - first_where
   - force_unwrapping
   - identical_operands
