+++ iOSTemplates/AppDelegate.swift
@@ -2,6 +2,7 @@
 //  AppDelegate.swift
 //
 //  Created by Khoa Vo T.A. VN.Danang on 23/12/2022.
+//  Copyright © 2022 Monstar Lab VietNam Co., Ltd. All rights reserved.
 //
 
 import IQKeyboardManagerSwift
