+++ iOSTemplates/View/Controller/Example/ExampleVC.xib
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="21507" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_12" orientation="portrait" appearance="light"/>
+    <device id="retina6_5" orientation="portrait" appearance="light"/>
     <dependencies>
         <deployment identifier="iOS"/>
         <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21505"/>
@@ -11,19 +11,47 @@
     <objects>
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="ExampleVC" customModule="iOSTemplates" customModuleProvider="target">
             <connections>
+                <outlet property="playButton" destination="dhj-XO-Egk" id="RSC-ad-JPa"/>
+                <outlet property="tableView" destination="ZtN-FL-cx2" id="qZA-nB-3VN"/>
                 <outlet property="view" destination="i5M-Pr-FkT" id="sfx-zR-JGt"/>
             </connections>
         </placeholder>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
         <view clearsContextBeforeDrawing="NO" contentMode="scaleToFill" id="i5M-Pr-FkT">
-            <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
+            <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="-1" estimatedSectionHeaderHeight="-1" sectionFooterHeight="-1" estimatedSectionFooterHeight="-1" translatesAutoresizingMaskIntoConstraints="NO" id="ZtN-FL-cx2">
+                    <rect key="frame" x="0.0" y="44" width="414" height="818"/>
+                    <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                </tableView>
+                <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="dhj-XO-Egk">
+                    <rect key="frame" x="334" y="782" width="50" height="50"/>
+                    <color key="backgroundColor" red="0.054901960780000002" green="0.68627450980000004" blue="0.10980392160000001" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+                    <constraints>
+                        <constraint firstAttribute="width" secondItem="dhj-XO-Egk" secondAttribute="height" multiplier="1:1" id="PFi-ub-isW"/>
+                        <constraint firstAttribute="width" constant="50" id="wtG-x5-LwA"/>
+                    </constraints>
+                    <color key="tintColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                    <state key="normal" title="Button"/>
+                    <buttonConfiguration key="configuration" style="plain" image="playstation.logo" catalog="system"/>
+                </button>
+            </subviews>
             <viewLayoutGuide key="safeArea" id="fnl-2z-Ty3"/>
             <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-            <point key="canvasLocation" x="42" y="21"/>
+            <constraints>
+                <constraint firstAttribute="trailing" secondItem="dhj-XO-Egk" secondAttribute="trailing" constant="30" id="2Fn-EH-0eU"/>
+                <constraint firstItem="ZtN-FL-cx2" firstAttribute="leading" secondItem="i5M-Pr-FkT" secondAttribute="leading" id="2vt-9I-Hw3"/>
+                <constraint firstItem="ZtN-FL-cx2" firstAttribute="top" secondItem="fnl-2z-Ty3" secondAttribute="top" id="E4b-3I-3vL"/>
+                <constraint firstItem="fnl-2z-Ty3" firstAttribute="bottom" secondItem="dhj-XO-Egk" secondAttribute="bottom" constant="30" id="MrF-Xq-MlG"/>
+                <constraint firstItem="fnl-2z-Ty3" firstAttribute="bottom" secondItem="ZtN-FL-cx2" secondAttribute="bottom" id="ZuG-X0-OIj"/>
+                <constraint firstAttribute="trailing" secondItem="ZtN-FL-cx2" secondAttribute="trailing" id="pGe-lE-oGG"/>
+            </constraints>
+            <point key="canvasLocation" x="45" y="21"/>
         </view>
     </objects>
     <resources>
+        <image name="playstation.logo" catalog="system" width="128" height="92"/>
         <systemColor name="systemBackgroundColor">
             <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
         </systemColor>
